refactor(theme): rebuild file structure
