+++ src/app/theme/customizer/customizer.component.scss
