+++ src/app/theme/header/branding.component.ts
