+++ src/app/theme/sidebar-notice/sidebar-notice.component.ts
