+++ src/app/theme/sidebar/sidebar.component.ts
