+++ src/app/theme/sidebar/user-panel.component.ts
