+++ src/app/theme/sidemenu/sidemenu.component.ts
