+++ src/app/theme/style/_core.scss
@@ -1,0 +1,6 @@
+@import 'core/reboot';
+@import 'core/branding';
+@import 'core/header';
+@import 'core/layout';
+@import 'core/sidenav';
+@import 'core/topnav';
