+++ src/app/theme/style/_mixins.scss
@@ -1,3 +1,4 @@
 @import 'mixins/badge';
 @import 'mixins/breakpoints';
+@import 'mixins/rtl';
 @import 'mixins/scrollbar';
