+++ src/app/theme/style/_widgets.scss
@@ -1,0 +1,2 @@
+@import 'widgets/badge';
+@import 'widgets/scrollbar';
