+++ src/app/theme/style/mixins/_rtl.scss
@@ -1,0 +1,5 @@
+@mixin rtl {
+  @at-root [dir='rtl'] #{&} {
+    @content;
+  }
+}
